cast to int in score

+++ src/mainApp/ScoreRecorder.java
@@ -25,8 +25,8 @@ public class ScoreRecorder {
     public void drawOn(Graphics2D g2){
         g2.drawString(  "score:"+score+"  "+
                         "level:"+level+"  "+
-                        "distance:"+(-GameViewer.getDistance())+"  "+
-                        "speed: "+(-GameViewer.getGameSpeed())+"   "+
+                        "distance:"+(int)(-GameViewer.getDistance())+"  "+
+                        "speed: "+(int)(-GameViewer.getGameSpeed())+"   "+
                         "numberCoins:"+numberCoins+"  ", 
                         50, 20);
